BACKEND/GrowthFarmAPI/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "growth-farm-api"
version = "1.0.0"
description = "A FastAPI backend for Growth Farm management"
authors = [{name = "Growth Farm Team"}]
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "pydantic[email]==2.5.0",
    "python-multipart==0.0.6",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-dotenv==1.0.0",
    "SQLAlchemy==1.4.53",
    "aiosqlite==0.19.0",
    "mysql-connector-python==8.2.0",
    "aiomysql==0.2.0",
    "pymysql==1.1.0",
    "email-validator==2.1.0",
    "google-generativeai==0.3.2",
]

[tool.fastapi]
app = "app.main:app"
dev = {app = "app.main:app"}
